# specs/002-fullstack-web-app/contracts/api-spec.yaml
openapi: 3.0.3
info:
  title: Todo Full-Stack Web Application API
  description: |
    RESTful API for multi-user authenticated todo application.
    All task endpoints require JWT authentication.
  version: 1.0.0
  contact:
    name: API Support
    url: https://github.com/user/todo-app

servers:
  - url: http://localhost:8000
    description: Local development server
  - url: http://backend:8000
    description: Docker Compose internal

security:
  - BearerAuth: []

paths:
  /health:
    get:
      summary: Health check endpoint
      description: Verify API is running (no authentication required)
      security: []
      responses:
        '200':
          description: API is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "healthy"

  /api/auth/register:
    post:
      summary: Register new user
      description: Create a new user account with email and password
      security: []
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegister'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Validation error (invalid email, weak password)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/auth/login:
    post:
      summary: User login
      description: Authenticate user and return JWT token
      security: []
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses:
        '200':
          description: Login successful, returns JWT token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/tasks:
    get:
      summary: List user's tasks
      description: Get all tasks belonging to authenticated user
      tags:
        - Tasks
      parameters:
        - name: is_completed
          in: query
          description: Filter by completion status
          schema:
            type: boolean
        - name: limit
          in: query
          description: Maximum number of tasks to return
          schema:
            type: integer
            default: 100
        - name: offset
          in: query
          description: Number of tasks to skip (pagination)
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                type: object
                properties:
                  tasks:
                    type: array
                    items:
                      $ref: '#/components/schemas/TaskResponse'
                  total:
                    type: integer
                    description: Total number of tasks (for pagination)
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    post:
      summary: Create new task
      description: Create a new task for authenticated user
      tags:
        - Tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Validation error (empty title, title too long)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  /api/tasks/{task_id}:
    get:
      summary: Get task details
      description: Retrieve a single task by ID (must belong to authenticated user)
      tags:
        - Tasks
      parameters:
        - name: task_id
          in: path
          required: true
          description: Task UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Forbidden - task belongs to another user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: Update task
      description: Update task title and/or description (must belong to authenticated user)
      tags:
        - Tasks
      parameters:
        - name: task_id
          in: path
          required: true
          description: Task UUID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Forbidden - task belongs to another user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Delete task
      description: Permanently delete a task (must belong to authenticated user)
      tags:
        - Tasks
      parameters:
        - name: task_id
          in: path
          required: true
          description: Task UUID
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Task deleted successfully (no content)
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Forbidden - task belongs to another user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/tasks/{task_id}/complete:
    patch:
      summary: Toggle task completion
      description: Toggle task between completed and incomplete status
      tags:
        - Tasks
      parameters:
        - name: task_id
          in: path
          required: true
          description: Task UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Task completion status toggled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Forbidden - task belongs to another user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/auth/login

  responses:
    UnauthorizedError:
      description: Authentication required or token invalid/expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Invalid authentication credentials"

  schemas:
    UserRegister:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          maxLength: 255
          example: "user@example.com"
        password:
          type: string
          format: password
          minLength: 8
          example: "SecurePass123"
          description: "Minimum 8 characters with at least one letter and one number"

    UserLogin:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          format: password
          example: "SecurePass123"

    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        email:
          type: string
          format: email
          example: "user@example.com"
        created_at:
          type: string
          format: date-time
          example: "2026-01-03T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2026-01-03T10:00:00Z"

    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        token_type:
          type: string
          example: "bearer"
        expires_in:
          type: integer
          description: "Token expiration in seconds"
          example: 604800

    TaskCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          maxLength: 200
          example: "Complete project documentation"
        description:
          type: string
          nullable: true
          example: "Write comprehensive docs for the full-stack app"

    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
          nullable: true
          example: "Updated task title"
        description:
          type: string
          nullable: true
          example: "Updated description"

    TaskResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "660e8400-e29b-41d4-a716-446655440001"
        user_id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        title:
          type: string
          example: "Complete project documentation"
        description:
          type: string
          nullable: true
          example: "Write comprehensive docs for the full-stack app"
        is_completed:
          type: boolean
          example: false
        created_at:
          type: string
          format: date-time
          example: "2026-01-03T11:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2026-01-03T11:00:00Z"

    Error:
      type: object
      properties:
        detail:
          type: string
          example: "Error description"
        field:
          type: string
          nullable: true
          example: "email"
          description: "Field name for validation errors"

tags:
  - name: Authentication
    description: User registration and login
  - name: Tasks
    description: Task CRUD operations (requires authentication)